third_party/WebKit/LayoutTests/http/tests/performance-timing/paint-timing/first-contentful-canvas.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FCP due to canvas</title>
</head>
<body>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <canvas id="canvas" width="200" height="200" ></canvas>

    <script>
    async_test(function (t) {
        const c = document.getElementById("canvas");
        const ctx = c.getContext("2d");
        ctx.beginPath();
        ctx.moveTo(0,0);
        ctx.lineTo(300,150);
        ctx.stroke();

        testRunner.capturePixelsAsyncThen(t.step_func_done(function() {
            const bufferedEntries = performance.getEntriesByType('paint');
            assert_equals(bufferedEntries.length, 2, "There should be two paint timing instances.");
            assert_equals(bufferedEntries[0].entryType, "paint");
            assert_equals(bufferedEntries[0].name, "first-paint");
            assert_equals(bufferedEntries[1].entryType, "paint");
            assert_equals(bufferedEntries[1].name, "first-contentful-paint");
        }));

    }, "First contentful paint fires due to canvas render.");
    </script>
</body>
</html>
